Revert "refactor: gameplay vote domain divide (#99)"
This reverts commit 753b94e616698c6bdefc935869399c33a583a345.
