--- frontend/src/features/gameplay/composition/index.ts
@@ -1,2 +1,0 @@
-export { useGameplayComposition } from "./hooks/useGameplayComposition";
-export { default as GameplayScreen } from "./components/GameplayScreen";
