+++ frontend/src/features/gameplay/session/hooks/useGameplayBootstrap.ts
@@ -1,6 +1,19 @@
 import { useCallback } from "react";
-import { sessionApi } from "../api/session.api";
-import { SessionBootstrapResult } from "../model/session.types";
+import { sessionApi, type RoundStateResponse } from "../api/session.api";
+import { RoundInfoState, SessionBootstrapResult } from "../model/session.types";
+
+function formatClockTime(date: Date): string {
+  const hours = String(date.getHours()).padStart(2, "0");
+  const minutes = String(date.getMinutes()).padStart(2, "0");
+  return `${hours}:${minutes}`;
+}
+
+function formatDuration(seconds: number): string {
+  const safeSeconds = Math.max(0, seconds);
+  const minutes = Math.floor(safeSeconds / 60);
+  const secs = safeSeconds % 60;
+  return `${minutes}:${String(secs).padStart(2, "0")}`;
+}
 
 export function useGameplayBootstrap() {
   const bootstrap = useCallback(async (): Promise<SessionBootstrapResult> => {
@@ -8,7 +21,22 @@ export function useGameplayBootstrap() {
     const { canvas, cells } = data;
 
     const roundRes = await sessionApi.getActiveRound(canvas.id);
-    const roundState = roundRes.data;
+    const roundState: RoundStateResponse = roundRes.data;
+
+    const round: RoundInfoState = {
+      roundId: roundState?.round?.id ?? null,
+      roundNumber: roundState?.round?.roundNumber ?? null,
+      roundDurationSec: roundState?.round?.roundDurationSec ?? null,
+      totalRounds: roundState?.round?.totalRounds ?? 0,
+      formattedGameEndTime: roundState?.round
+        ? formatClockTime(new Date(roundState.round.gameEndAt))
+        : null,
+      remainingSeconds: roundState?.timer?.remainingSeconds ?? null,
+      formattedRemainingTime: roundState?.timer
+        ? formatDuration(roundState.timer.remainingSeconds)
+        : null,
+      isRoundExpired: roundState?.timer?.isRoundExpired ?? false,
+    };
 
     let remaining: number | null = null;
 
@@ -22,8 +50,7 @@ export function useGameplayBootstrap() {
       gridX: canvas.gridX,
       gridY: canvas.gridY,
       cells,
-      round: roundState?.round ?? null,
-      timer: roundState?.timer ?? null,
+      round,
       remaining,
     };
   }, []);
