--- frontend/src/features/gameplay/vote/index.ts
@@ -1,15 +1,0 @@
-export { gameplayVoteApi } from "./api/vote.api";
-
-export { useVoteState } from "./hooks/useVoteState";
-export { useVotePopupState } from "./hooks/useVotePopupState";
-
-export { default as VotePanel } from "./components/VotePanel";
-export { default as VoteTickets } from "./components/VoteTickets";
-export { default as VoteStatusBoard } from "./components/VoteStatusBoard";
-export { default as VotePopup } from "./components/VotePopup";
-export { default as VoteResultList } from "./components/VoteResultList";
-export { default as ColorPalette } from "./components/ColorPalette";
-export { default as ColorSlotGrid } from "./components/ColorSlotGrid";
-
-export * from "./model/vote.types";
-export * from "./model/vote.selectors";
